extract parse code to file, refactor magic numbers
