+++ src/ardor.h
@@ -20,23 +20,40 @@
 #include "ux.h"
 #include "cx.h"
 #include "os.h"
+#include "config.h"
 
 #if defined(TARGET_NANOS)
     unsigned int makeTextGoAround_preprocessor(bagl_element_t * const element);
 #endif
 
 uint64_t publicKeyToId(const uint8_t * const publicKey);
-uint8_t ardorKeys(const uint8_t * const derivationPath, const uint8_t derivationPathLengthInUints32, 
+uint8_t ardorKeys(const uint8_t * const derivationPath, const uint8_t derivationPathLengthInUints32,
                     uint8_t * const keySeedBfrOut, uint8_t * const publicKeyCurveXout, uint8_t * const publicKeyEd25519YLEWithXParityOut, uint8_t * const chainCodeOut, uint16_t * const exceptionOut);
+char * chainName(const uint8_t chainId);
 
 void signMsg(uint8_t * const keySeedBfr, const uint8_t * const msgSha256, uint8_t * const sig);
 
 void ui_idle();
 bool check_canary();
 
-uint8_t getSharedEncryptionKey(const uint8_t * const derivationPath, const uint8_t derivationPathLengthInUints32, const uint8_t* const targetPublicKey, 
+uint8_t getSharedEncryptionKey(const uint8_t * const derivationPath, const uint8_t derivationPathLengthInUints32, const uint8_t* const targetPublicKey,
                                 const uint8_t * const nonce, uint16_t * const exceptionOut, uint8_t * const aesKeyOut);
 
+//the amount of digits on the right of the decimal dot for each chain
+uint8_t chainNumDecimalsBeforePoint(const uint8_t chainId);
+
+uint8_t formatAmount(char * const outputString, const uint16_t maxOutputLength, uint64_t numberToFormat, const uint8_t numDigitsBeforeDecimal);
+
+// define max text sizes for the different UI screens
+#define MAX_FEE_TEXT_SIZE 21                //9,223,372,036,854,775,807 is the biggest number you can hold in uint64 + the dot + null
+#define MAX_CHAIN_AND_TXN_TYPE_TEXT_SIZE 60 //Aproximation of size
+#define MAX_WIN1_TITLE_SIZE 9               //MAX("Amount","Asset Id")
+#define MAX_WIN1_TEXT_SIZE 31               //same as fee text + name of the chain + space
+#define MAX_WIN2_TITLE_SIZE 20              //The longest string is price per (some chain name here)
+#define MAX_WIN2_TEXT_SIZE 31               //MAX(Ardor arddress = 27, feeText + chainName)
+#define MAX_WIN3_TITLE_SIZE 10              //MAX("Recipient")
+#define MAX_WIN3_TEXT_SIZE 28               //MAX(Ardor arddress = 27)
+#define MAX_APPENDAGES_TEXT_SIZE 60         //this should allow displaying the names for up to three types, otherwise we show a bitmap
 
 //This is the state object that authAndSignTxn uses
 typedef struct {
@@ -73,15 +90,15 @@ typedef struct {
    	uint16_t txnSizeBytes;                                 //The decalred Txn size
 
 
-    char feeText[21];               //9,223,372,036,854,775,807 is the biggest number you can hold in uint64 + the dot + null terminator means the longest text is 20
-    char chainAndTxnTypeText[60];   //Aproximation of size
-    char optionalWindow1Title[9];   //MAX("Amount","Asset Id")
-    char optionalWindow1Text[31];   //same as fee text + name of the chain + space
-    char optionalWindow2Title[20];  //The longest string is price per (some chain name  here)
-    char optionalWindow2Text[31];   //MAX(Ardor arddress = 27, feeText + chainName)
-    char optionalWindow3Title[10];  //MAX("Recipient")
-    char optionalWindow3Text[28];   //MAX(Ardor arddress = 27)
-    char appendagesText[60];        //this should allow displaying the names for up to three types, otherwise we show a bitmap
+    char feeText[MAX_FEE_TEXT_SIZE];
+    char chainAndTxnTypeText[MAX_CHAIN_AND_TXN_TYPE_TEXT_SIZE];
+    char optionalWindow1Title[MAX_WIN1_TITLE_SIZE];
+    char optionalWindow1Text[MAX_WIN1_TEXT_SIZE];
+    char optionalWindow2Title[MAX_WIN2_TITLE_SIZE];
+    char optionalWindow2Text[MAX_WIN2_TEXT_SIZE];
+    char optionalWindow3Title[MAX_WIN3_TITLE_SIZE];
+    char optionalWindow3Text[MAX_WIN3_TEXT_SIZE];
+    char appendagesText[MAX_APPENDAGES_TEXT_SIZE];
     uint8_t uiFlowBitfeild;         //This is a bit feild for selecting the right UI flow
 
 } authTxn_t;
@@ -89,7 +106,7 @@ typedef struct {
 //State for the encryptDecrypt handler
 typedef struct {
     uint8_t mode;                           //Modes are described in the .C file
-    uint8_t cbc[16];                        //Something to do with AES state
+    uint8_t cbc[CX_AES_BLOCK_SIZE];         //Something to do with AES state
     cx_aes_key_t aesKey;                    //This is the encryption key
 } encyptionState_t;
